2021 day 8: part 2 (+ redo)
I gave up on Clojure. This was already a fairly hard problem, even in
PHP it took me a while. I just could not get all the right loops over
the data in Clojure... since it was my first day of working with it
I had to look up everything. It took me hours.

Luckily I could do fancy in PHP also.